+++ src/Commands/Console/SeedData.php
@@ -3,12 +3,11 @@
 namespace Webkul\DataFaker\Commands\Console;
 
 use DB;
+use Faker\Generator as Faker;
 use Illuminate\Console\Command;
 use Webkul\DataFaker\Database\Seeders\CustomerTableDataSeeder;
 use Webkul\DataFaker\Database\Seeders\CategoryTableDataSeeder;
-// use Webkul\DataFaker\Database\Seeders\ProductTableDataSeeder;
-
-use Faker\Generator as Faker;
+use Webkul\DataFaker\Database\Seeders\ProductTableDataSeeder;
 use Webkul\Category\Repositories\CategoryRepository;
 
 class SeedData extends Command
@@ -70,6 +69,8 @@ class SeedData extends Command
                             DB::table('addresses')->where('address_type', 'customer')->delete();
                         }
 
+                        $this->comment('Seeding...');
+
                         $seeder = new CustomerTableDataSeeder;
                         $seeder->callWith(CustomerTableDataSeeder::class, [$count]);
 
@@ -79,6 +80,17 @@ class SeedData extends Command
                         if ($this->confirm('Do you want to seed product category?')) {
                             session()->put('seed_product_category', true);
                         }
+
+                        if ($this->confirm('Do you want to seed configurable products ?')) {
+                            session()->put('seed_config_product', true);
+                        }
+
+                        $this->comment('Seeding...');
+
+                        $seeder = new ProductTableDataSeeder;
+                        $seeder->callWith(ProductTableDataSeeder::class, [$count]);
+
+                        $this->comment('Products Created Successfully.');
                         break;
                     case 3:
                         $parent = $this->ask('Enter the number of parent categories');
@@ -94,8 +106,9 @@ class SeedData extends Command
 
                         if($parent > 0 && $child > 0) {
 
-                            $categorySeeder = new CategoryTableDataSeeder();
+                            $this->comment('Seeding...');
 
+                            $categorySeeder = new CategoryTableDataSeeder();
                             $categorySeeder->callWith(CategoryTableDataSeeder::class, [$parent, $child,$this->faker]);
 
                             $this->comment('Categories Created Successfully.');
@@ -111,94 +124,7 @@ class SeedData extends Command
             $this->comment('Warning: Please select a valid number.');
         }
 
-dd('test');
-        if ($this->confirm('Confirm: Please confirm, Do you want to remove all tables and records?')) {
-            $migrate = shell_exec('php artisan migrate:fresh');
-        } else {
-            $migrate = shell_exec('php artisan migrate');
-        }
-
-        $this->info($migrate);
-
-        // running `php artisan vendor:publish --all`
-        $this->warn('Step: Publishing Assets and Configurations...');
-        $result = shell_exec('php artisan vendor:publish --all');
-        $this->info($result);
-
-        // running `php artisan storage:link`
-        $this->warn('Step: Linking Storage directory...');
-        $result = shell_exec('php artisan storage:link');
-        $this->info($result);
-
-        $this->comment('Info: Generating super user for the system.');
-        $name = $this->ask('Input: Please enter super user name?');
-
-        // $validator = Validator::make([
-        //     'first_name' => $name
-        // ], [
-        //     'first_name' => 'required',
-        // ]);
-
-        // if ($validator->fails()) {
-        //     $this->comment('Warning: Name invalid, please enter try again.');
-
-        //     return false;
-        // }
-
-        // $this->comment('Info: You entered = '. $name);
-        // $email = $this->ask('Input: Please enter email?');
-
-        // $data = [
-        //     'email' => $email
-        // ];
-
-        // $validator = Validator::make($data, [
-        //     'email' => 'required|email|unique:super_admins,email',
-        // ]);
-
-        // if($validator->fails()) {
-        //     $this->comment('Warning: Email already exists or invalid, please enter try again.');
-
-        //     return false;
-        // }
-
-        // unset($data);
-
-        // $this->comment('Info: You entered = ' . $email);
-        // $password = $this->ask('Input: Please enter password?');
-        // $data = ['password' => $password];
-
-        // $validator = Validator::make($data, [
-        //     'password' => 'required|string|min:6'
-        // ]);
-
-        // if ($validator->fails()) {
-        //     $this->comment('Warning: Password invalid, make sure password is atleast 6 characters of length.');
-
-        //     return false;
-        // }
-
-        // $this->comment('Info: You entered = '. $password);
-
-        // unset($data);
-
-        // if ($this->confirm('Confirm: Please confirm all the entered details are correct?')) {
-        //     $data = [
-        //         'first_name'    => $name,
-        //         'email'         => $email,
-        //         'password'      => bcrypt($password),
-        //     ];
-
-        //     $result = $this->generateSuperUserCompany($data);
-
-        //     if ($result) {
-        //         $this->comment('Success: Super user for the system is created successfully.');
-        //     } else {
-        //         $this->comment('Warning: Super user for the system already exists, please contact support@bagisto.com for troubleshooting.');
-        //     }
-
-        // } else {
-        //     $this->comment('Warning: Please try again for creating the super user.');
-        // }
+        session()->forget('seed_product_category');
+        session()->forget('seed_config_product');
     }
 }