+++ src/Database/Factories/Product/ProductInventoryFactory.php
@@ -2,9 +2,8 @@
 
 namespace Webkul\DataFaker\Database\Factories\Product;
 
-use Illuminate\Database\Eloquent\Factories\Factory;
-use Faker\Generator as Faker;
 use Webkul\Product\Models\ProductInventory;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
 class ProductInventoryFactory extends Factory
 {