+++ src/Database/Factories/ProductAttributeValueFactory.php
@@ -1,7 +1,0 @@
-<?php
-
-use Faker\Generator as Faker;
-
-$factory->define(\Webkul\Product\Models\ProductAttributeValue::class, function (Faker $faker, $attributeValue) {
-    return $attributeValue;
-});