+++ src/Database/Factories/ProductImageFactory.php
@@ -1,7 +1,0 @@
-<?php
-
-use Faker\Generator as Faker;
-
-$factory->define(\Webkul\Product\Models\ProductImage::class, function ($faker,$fakeImage) {
-    return $fakeImage;
-});