+++ CHANGELOG.md
@@ -1,6 +1,22 @@
-# CHANGELOG for v1.2.x
+# CHANGELOG for v1.3.x
 
 #### This changelog consists the bug & security fixes and new features being included in the releases listed below.
+
+## **v1.3.3(24-Jan-2022)** - *Release*
+
+* [compatibility] Compatible with bagisto v1.3.3.
+
+* [compatibility] Seed fake data using Laravel 8 Modal Factory.
+
+* [feature] Seed the data easily form the terminal.
+
+* [feature] Seed the customers.
+
+* [feature] Seed the categories.
+
+* [feature] Seed the products[simple and configurable].
+
+* [feature] Seed the product category.
 
 ## **v1.2.1(9-10-2020)** - *Release*